requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
cellpose==2.1.1
    # via -r requirements.in
django==4.1.5
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
fastremap==1.13.3
    # via cellpose
imagecodecs==2022.12.24
    # via cellpose
llvmlite==0.39.1
    # via
    #   cellpose
    #   numba
natsort==8.2.0
    # via cellpose
numba==0.56.4
    # via cellpose
numpy==1.23.5
    # via
    #   cellpose
    #   fastremap
    #   imagecodecs
    #   numba
    #   opencv-python
    #   opencv-python-headless
    #   scipy
    #   tifffile
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
opencv-python==4.7.0.68
    # via -r requirements.in
opencv-python-headless==4.7.0.68
    # via cellpose
pillow==9.4.0
    # via -r requirements.in
scipy==1.10.0
    # via cellpose
sqlparse==0.4.3
    # via django
tifffile==2022.10.10
    # via cellpose
torch==1.13.1
    # via cellpose
tqdm==4.64.1
    # via cellpose
typing-extensions==4.4.0
    # via torch
wheel==0.38.4
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11

# The following packages are considered to be unsafe in a requirements file:
# setuptools
